Added OrderArchive entity, Adress entity + changes

+++ src/main/java/com/example/projectMara/domain/model/Movie.java
@@ -32,7 +32,7 @@ public class Movie {
     private LocalDateTime createdAt;
 
     @Column(name = "avg_rate")
-    private float avgRate;
+    private double avgRate;
 
     @Column(name = "premiere_date")
     private LocalDate premiereDate;
